login functionality and restrictions

+++ base/views.py
@@ -1,5 +1,9 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 from django.http import HttpResponse
 from .models import Room, Topic
 from .forms import RoomForm
@@ -17,11 +21,50 @@ def safety(request):
     return render(request,'navbar_links/safety.html')
 def support(request):
     return render(request,'navbar_links/support.html')
-def login(request):
-    return render(request,'navbar_links/login.html')
+
+
+
+
+
+
 
 
 #These are the main webpages of the studybud
+
+def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request,'User does not exist')
+
+        user = authenticate(request,username=username,password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+       
+       
+    context ={}
+    return render (request,'base/login.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+
+
+
+
 def home(request):
 
     q = request.GET.get('q') if  request.GET.get('q') != None else ''
@@ -44,6 +87,10 @@ def room(request, pk):
     return render(request,'base/room.html', context)
 
 
+
+
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -57,9 +104,14 @@ def createRoom(request):
     return render(request,'base/room_form.html',context )
 
 
+
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Must have created the post to edit ')
 
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
@@ -71,8 +123,16 @@ def updateRoom(request,pk):
     return render(request,'base/room_form.html',context)
 
 
+
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+
+    if request.user != room.user:
+        return HttpResponse('Must have created the post to delete ')
+    
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
